Added support to exclude http client
Fixes #141
