+++ logbook-spring-boot-starter/src/test/java/org/zalando/logbook/spring/Application.java
@@ -3,6 +3,11 @@ package org.zalando.logbook.spring;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 
-@ImportAutoConfiguration({LogbookAutoConfiguration.class, JacksonAutoConfiguration.class})
+@ImportAutoConfiguration({
+        LogbookAutoConfiguration.class,
+        LogbookSecurityAutoConfiguration.class,
+        LogbookHttpClientAutoConfiguration.class,
+        JacksonAutoConfiguration.class
+})
 public class Application {
 }
